Co-locate some test files
