+++ src/pages/localStorage.test.ts
@@ -1,9 +1,9 @@
 import { compressToUTF16 } from 'lz-string';
 import { describe, expect, test } from 'vitest';
 
-import { defaultState } from '../../commons/application/ApplicationTypes';
-import type { SavedState } from '../localStorage';
-import { loadStoredState, saveState } from '../localStorage';
+import { defaultState } from '../commons/application/ApplicationTypes';
+import type { SavedState } from './localStorage';
+import { loadStoredState, saveState } from './localStorage';
 
 const mockShortDefaultState: SavedState = {
   session: {
